% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullR.plotly-package.R
\docType{package}
\name{WeibullR.plotly-package}
\alias{WeibullR.plotly}
\alias{WeibullR.plotly-package}
\title{WeibullR.plotly: Interactive Weibull Probability Plots}
\description{
Build interactive Weibull Probability Plots with 'WeibullR' by David Silkworth and Jurgen Symynck (2022) \url{https://CRAN.R-project.org/package=WeibullR}, an R package for Weibull analysis, and 'plotly' by Carson Sievert (2020) \url{https://plotly-r.com}, an interactive web-based graphing library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/WeibullR.plotly/}
  \item \url{https://github.com/paulgovan/WeibullR.plotly}
  \item Report bugs at \url{https://github.com/paulgovan/WeibullR.plotly/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
